+++ dstep/core/Core.d
@@ -28,3 +28,24 @@ T memoize(T)(ref Optional!T cache, scope T delegate() operation)
 {
     return cache.or(operation().tap!((T e) => cache = e));
 }
+
+template flatMap(func...)
+if (func.length >= 1)
+{
+    import std.range : isInputRange;
+    import std.traits : Unqual;
+    import std.algorithm : cache, map, joiner;
+
+    auto flatMap(Range)(Range range)
+    if (isInputRange!(Unqual!Range)) => range.map!func.cache.joiner;
+}
+
+///
+unittest
+{
+    import std.algorithm : equal;
+
+    [1, 2, 3, 4]
+        .flatMap!(e => [e, -e])
+        .equal([1, -1, 2, -2, 3, -3, 4, -4]);
+}